requirements.txt
django-cors-headers==3.10.1
  Django==4.0
    asgiref==3.4.1
    backports.zoneinfo==0.2.1
    sqlparse==0.4.2
djangorestframework-simplejwt==5.0.0
  Django==4.0
    asgiref==3.4.1
    backports.zoneinfo==0.2.1
    sqlparse==0.4.2
  djangorestframework==3.13.1
    Django==4.0
      asgiref==3.4.1
      backports.zoneinfo==0.2.1
      sqlparse==0.4.2
    pytz==2021.3
  PyJWT==2.3.0
ffmpeg-python==0.2.0
  future==0.18.2
psycopg2==2.9.2
smart-getenv==1.1.0
wheel==0.37.1
